% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHTE.R
\name{eHTE}
\alias{eHTE}
\title{eHTE}
\usage{
eHTE(placebo_arm, drug_arm, pctiles = seq(from = 3, to = 97, by = 2)/100)
}
\arguments{
\item{placebo_arm}{A vector of patient-level treatment effects from the control condition (e.g., placebo).}

\item{drug_arm}{A vector of patient-level treatment effects from the active condition (e.g., active drug).}

\item{pctiles}{A vector of percentiles at which to evaluate the drug/placebo difference. Default is a vector of percentiles between the 3rd and 97th at increments of 2 (as in Siegel et al 2024).}
}
\value{
An custom object of class 'eHTE_class'.
}
\description{
Estimates the eHTE test statistic, and optionally the p-value associated with it, for two samples of possibly uneven length.
}
