% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raison_data.R
\docType{data}
\name{raison_data}
\alias{raison_data}
\title{Loebel et al 2013 Example Dataset}
\format{
A dataframe with 94 rows (each a distinct patient) and 5 columns:
\itemize{
\item{SUBJID}{Subject ID: Sub1, Sub2, etc}
\item{arm}{Arm of the study: Placebo, or Psilocybin}
\item{Baseline}{Baseline on the MADRS}
\item{Week 6}{Week 6 on the MADRS}
\item{MADRSchg}{Change from Baseline at Week 6 on the MADRS}
}
@section Source:
This data was retrieved from \href{https://gitlab.com/siegelandthebrain1/ehte/-/tree/main/eHTE_matlab_package?ref_type=heads}{Joshua Siegel's Gitlab Page for his eHTE package}.
}
\usage{
data(raison_data)
}
\description{
This is the patient-level change data originally reported in \href{https://jamanetwork.com/journals/jama/article-abstract/2808950}{Raison et al 2023 }.
}
