% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHTE.R
\name{testHTE}
\alias{testHTE}
\title{testHTE}
\usage{
testHTE(estimateHTE_result, n_perm = 100)
}
\arguments{
\item{estimateHTE_result}{An object of class eHTE_class - for example, the result of estimateHTE()}

\item{n_perm}{The number of permutations desired. Default = 100.}
}
\value{
A custom object of class 'eHTE_class', which accepts methods summary(), plot(), and print().
}
\description{
Compute a p-value of an obtained eHTE estimate, using non-parametric permutation.
}
