% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{eHTE}
\alias{eHTE}
\alias{eHTE-package}
\title{eHTE: A Package for Estimating Heterogeneous Treatment Effects}
\description{
The eHTE package provides tools for estimating heterogeneous treatment effects.
This work is based on Siegel et al, 2024: https://www.medrxiv.org/content/10.1101/2024.04.23.24306211v1
}
\section{Functions}{

The main functions in this package are:
\itemize{
\item \code{\link{estimateHTE}}: Estimates the eHTE metric of Siegel et al.
\item \code{\link{testHTE}}: Tests the eHTE metric using the non-parameteric permutation method described by Siegel et al.
}
}

\author{
Chris Chatham \email{chathach@gmail.com}
}
\keyword{package}
